# Avoid tracing allocations if tracing is disabled anyway.