<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

## Proposed Changes

If tracing is disabled we don't need to allocate spans and contexts at all. It saves 6 allocations in the activator per request. I'll post benchmark data once #6533 merges.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @vagababov