@markusthoemmes 2020-01-16
The promised benchmark results:

### Before

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-sequential-12         	  174910	      6416 ns/op	    3140 B/op	      36 allocs/op
BenchmarkHandler/002k-resp-len-parallel-12           	  987793	      2012 ns/op	    3081 B/op	      36 allocs/op
BenchmarkHandler/016k-resp-len-sequential-12         	  178777	     10411 ns/op	    3129 B/op	      36 allocs/op
BenchmarkHandler/016k-resp-len-parallel-12           	  836883	      1923 ns/op	    3188 B/op	      36 allocs/op
BenchmarkHandler/032k-resp-len-sequential-12         	  153043	      7984 ns/op	    3079 B/op	      36 allocs/op
BenchmarkHandler/032k-resp-len-parallel-12           	  652132	      2361 ns/op	    3281 B/op	      36 allocs/op
BenchmarkHandler/064k-resp-len-sequential-12         	  137919	      9287 ns/op	    3143 B/op	      36 allocs/op
BenchmarkHandler/064k-resp-len-parallel-12           	  648163	      2200 ns/op	    3212 B/op	      36 allocs/op
BenchmarkHandler/128k-resp-len-sequential-12         	  107845	     11807 ns/op	    3142 B/op	      36 allocs/op
BenchmarkHandler/128k-resp-len-parallel-12           	  599925	      2478 ns/op	    3230 B/op	      36 allocs/op
PASS
```

### After

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-sequential-12         	  173229	      5963 ns/op	    2761 B/op	      30 allocs/op
BenchmarkHandler/002k-resp-len-parallel-12           	 1000000	      1736 ns/op	    2690 B/op	      30 allocs/op
BenchmarkHandler/016k-resp-len-sequential-12         	  167443	      6477 ns/op	    2782 B/op	      30 allocs/op
BenchmarkHandler/016k-resp-len-parallel-12           	 1000000	      1441 ns/op	    2689 B/op	      30 allocs/op
BenchmarkHandler/032k-resp-len-sequential-12         	  176360	      7484 ns/op	    2749 B/op	      30 allocs/op
BenchmarkHandler/032k-resp-len-parallel-12           	 1000000	      1696 ns/op	    2690 B/op	      30 allocs/op
BenchmarkHandler/064k-resp-len-sequential-12         	  156718	      8084 ns/op	    2828 B/op	      30 allocs/op
BenchmarkHandler/064k-resp-len-parallel-12           	 1000000	      1850 ns/op	    2692 B/op	      30 allocs/op
BenchmarkHandler/128k-resp-len-sequential-12         	  108759	     10195 ns/op	    2750 B/op	      30 allocs/op
BenchmarkHandler/128k-resp-len-parallel-12           	  751768	      2358 ns/op	    2754 B/op	      30 allocs/op
PASS
```